Add probability tables for both monsters and code to take those tables and construct probability tables for all the monsters that could appear at the current depth which can be looped over to select monsters randomly. See below for more information.
Will keep items and monsters names as strings in a dictionary as keys mapping to a dictionary of ( K , V) pairs where K is a dungeon depth and V is the probability as a float x where 0.0<= x <= 1.0.
The table of monsters will be looped over and ones with keys corresponding to the current_depth (those who have a non-zero probability of appearing at the current depth) will have their probabilities extracted and will then be stored in a new dictionary whose keys are the monsters that could appear and the values are their relative probabilities of appearing , weighted such that the sum of such relative probabilities across all the possibly appearing monsters is 1.0. Then the table will be used to randomly select monsters (or equivalently but separately items) until a set maximum has been spawned.